feat(calendar): popup on click
Når man klikker på et arrangement i kalender vil det poppe opp en informasjonsboks.
Kalenderen er også lagt til i arrangementsiden.
